+++ api/pkg/db/models/rankingaverage.go
@@ -6,7 +6,7 @@ type RankingAverage struct {
 	WcaId   string     `json:"wcaId" gorm:"column:wca_id"`
 	StateId string     `json:"stateId" gorm:"column:state_id"`
 	EventId string     `json:"eventId" gorm:"column:event_id"`
-	Result  null.Int32 `json:"result" gorm:"column:average"`
+	Result  null.Int64 `json:"result" gorm:"column:average"`
 	Ranking int32      `json:"ranking" gorm:"column:ranking"`
 }
